Variables to remove static paths
I removed all static paths and turned them into variables. This means that the paths can be set once as a value for the variable and will be reused in their original places. 

The variable for the path was updated to use the default bash environment variable $USER instead of the hard-coded username.

The variable for the path is now read in from user input after prompts showing the default value. This could be improved to reduce recycled paths. 

The variable for the path will go to a default if it is not entered by the user. The default can be changed at the top of the script.

The variable for the path will be examined to see if it exists, otherwise the script will exit with a failure.